Add files via upload
A security incident has occurred, resulting in the unauthorized access and theft of valuable data from our fictional entity, Keiko Corp. The primary objective of this project is to meticulously analyze a substantial database, identifying potential perpetrators. This undertaking underscores proficiency in data filtering, ordering, and the adept utilization of table joins within the database. The ultimate resolution is derived through a meticulous cross-referencing process between two distinct databases, culminating in the retrieval of pertinent information.